backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
python-dotenv==1.1.0
httpx==0.28.1
chromadb==1.5.6
sentence-transformers==4.1.0
pypdf==5.4.0
loguru==0.7.3
numpy==2.4.4
